# [V] Use flattened slice