Main: `48.82 ms  |  743.93 ms |  6.60 s`

This PR:
```
Vlang:

        Benchmark 1: ./related
        Processing time (w/o IO): 23.351ms
        Processing time (w/o IO): 26.541ms
        Processing time (w/o IO): 24.216ms
        Processing time (w/o IO): 23.916ms
        Processing time (w/o IO): 23.535ms
        Processing time (w/o IO): 24.046ms
        Processing time (w/o IO): 23.858ms
        Processing time (w/o IO): 24.002ms
        Processing time (w/o IO): 23.984ms
        Processing time (w/o IO): 24.424ms
        Processing time (w/o IO): 23.227ms
        Processing time (w/o IO): 24.020ms
        Processing time (w/o IO): 23.963ms
          Time (mean ± σ):     328.5 ms ±   3.5 ms    [User: 305.5 ms, System: 22.7 ms]
          Range (min … max):   323.2 ms … 333.3 ms    10 runs
         
Vlang:

        Benchmark 1: ./related
        Processing time (w/o IO): 357.253ms
        Processing time (w/o IO): 355.159ms
        Processing time (w/o IO): 355.669ms
          Time (mean ± σ):      7.326 s ±  0.198 s    [User: 7.212 s, System: 0.112 s]
          Range (min … max):    7.186 s …  7.466 s    2 runs
         
Vlang:

        Benchmark 1: ./related
        Processing time (w/o IO): 3.136s
        Processing time (w/o IO): 3.133s
        Processing time (w/o IO): 3.151s
          Time (mean ± σ):     70.916 s ±  0.283 s    [User: 70.547 s, System: 0.361 s]
          Range (min … max):   70.716 s … 71.116 s    2 runs
```